+++ xmake/templates/c++/static_library_tbox/project/src/_library/xmake.lua
@@ -8,8 +8,7 @@ target("[targetname]")
     add_defines("__tb_prefix__=\"[targetname]\"")
 
     -- set the auto-generated config.h
-    set_config_h("$(buildir)/[targetname]/config.h")
-    set_config_h_prefix("CONFIG")
+    set_config_header("$(buildir)/[targetname]/config.h", {prefix = "CONFIG"})
 
     -- add the header files for installing
     add_headers("../([targetname]/**.h)")
@@ -18,11 +17,8 @@ target("[targetname]")
     add_includedirs("$(buildir)")
     add_includedirs("$(buildir)/[targetname]")
 
-    -- set the object files directory
-    set_objectdir("$(buildir)/.objs")
-
     -- add packages
-    add_options("tbox", "base")
+    add_packages("tboox.tbox", "base")
 
     -- add files
     add_files("*.cpp")