+++ xmake/templates/c/console_tbox/project/src/xmake.lua
@@ -8,14 +8,10 @@ target("[targetname]")
     add_defines("__tb_prefix__=\"[targetname]\"")
 
     -- set the auto-generated config.h
-    set_config_h("$(buildir)/[targetname].config.h")
-    set_config_h_prefix("CONFIG")
-
-    -- set the object files directory
-    set_objectdir("$(buildir)/.objs")
+    set_config_header("$(buildir)/[targetname].config.h", {prefix = "CONFIG"})
 
     -- add packages
-    add_options("tbox", "base")
+    add_packages("tboox.tbox", "base")
 
     -- add files
     add_files("*.c")